+++ src/index.tsx
@@ -12,4 +12,6 @@ export * from './modules/pass';
 
 export * from './modules/checkout';
 
+export * from './modules/dashboard';
 export * from './modules/addFunds';
+export * from './modules/poll';