+++ src/modules/auth/components/CompleteSignUpSuccess/CompleteSignUpSuccess.tsx
@@ -2,8 +2,9 @@ import classNames from 'classnames';
 
 import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
 import { useCompanyConfig } from '../../../shared/hooks/useCompanyConfig';
-import useRouter from '../../../shared/hooks/useRouter';
+import { useRouterConnect } from '../../../shared/hooks/useRouterConnect';
 import useTranslation from '../../../shared/hooks/useTranslation';
+import { usePixwayAuthentication } from '../../hooks/usePixwayAuthentication';
 import { AuthButton } from '../AuthButton';
 import { AuthFooter } from '../AuthFooter';
 import { AuthLayoutBase, AuthLayoutBaseClasses } from '../AuthLayoutBase';
@@ -15,7 +16,14 @@ interface Props {
 export const CompleteSignUpSuccess = ({ classes = {} }: Props) => {
   const { logoUrl } = useCompanyConfig();
   const [translate] = useTranslation();
-  const router = useRouter();
+  const router = useRouterConnect();
+  const { signOut } = usePixwayAuthentication();
+
+  const handleContinue = () => {
+    signOut().then(() => {
+      router.pushConnect(PixwayAppRoutes.SIGN_IN);
+    });
+  };
 
   return (
     <AuthLayoutBase
@@ -31,17 +39,13 @@ export const CompleteSignUpSuccess = ({ classes = {} }: Props) => {
     >
       <div className="pw-text-[#35394C] sm:!pw-mt-6">
         <h1 className="pw-mb-6 pw-text-center pw-text-2xl pw-leading-[29px] pw-font-bold px- sm:pw-px-0">
-          Conta criada com sucesso!
+          {translate('companyAuth>signUp>accountCreated')}
         </h1>
         <p className="pw-text-center pw-mb-6 pw-text-sm sm:pw-text-[13px] pw-leading-[15px] sm:pw-leading-[17px] pw-px-[70px] sm:pw-px-0">
           {translate('auth>completeSignUp>mailAlreadyVerified')}
         </p>
 
-        <AuthButton
-          fullWidth
-          className="pw-mb-6"
-          onClick={() => router.push(PixwayAppRoutes.HOME)}
-        >
+        <AuthButton fullWidth className="pw-mb-6" onClick={handleContinue}>
           {translate('loginPage>formSubmitButton>signIn')}
         </AuthButton>
         <AuthFooter />