+++ src/modules/index.ts
@@ -3,5 +3,7 @@ export * from './core';
 export * from './auth';
 export * from './tokens';
 export * from './checkout';
+export * from './dashboard';
 export * from './addFunds';
 export * from './pass';
+export * from './poll';