+++ src/modules/shared/components/HeaderPixwaySDK/components/NavigationTabsPixwaySDK/components/NavigationTabsPixwaySDKMobile.tsx
@@ -4,7 +4,7 @@ import { ReactComponent as CloseIcon } from '../../../../../assets/icons/closeIc
 import { ReactComponent as HamburguerIcon } from '../../../../../assets/icons/headerHamburger.svg';
 import { PixwayAppRoutes } from '../../../../../enums/PixwayAppRoutes';
 import { usePixwaySession } from '../../../../../hooks/usePixwaySession';
-import useRouter from '../../../../../hooks/useRouter';
+import { useRouterConnect } from '../../../../../hooks/useRouterConnect';
 import useTranslation from '../../../../../hooks/useTranslation';
 import { PixwayButton } from '../../../../PixwayButton/PixwayButton';
 import { NavigationTabsPixwaySDKProps } from '../NavigationTabsPixwaySDK';
@@ -16,9 +16,11 @@ export const NavigationTabsPixwaySDKMobile = ({
   toogleMenu,
   signInRoute = PixwayAppRoutes.SIGN_IN,
   signUpRoute = PixwayAppRoutes.SIGN_UP,
+  textColor = 'black',
+  hasSignUp,
 }: NavigationTabsPixwaySDKProps) => {
   const [translate] = useTranslation();
-  const router = useRouter();
+  const router = useRouterConnect();
   const [openedTabs, setOpenedTabs] = useState<boolean>(false);
   const { data: session } = usePixwaySession();
 
@@ -28,20 +30,26 @@ export const NavigationTabsPixwaySDKMobile = ({
     } else setOpenedTabs(!openedTabs);
   };
 
-  return (
+  return !session || (tabs && tabs.length > 0) ? (
     <div className={` ${classNames?.className}`}>
       {opened ? (
-        <CloseIcon className="pw-cursor-pointer" onClick={toggleTabsMemo} />
+        <CloseIcon
+          style={{ stroke: textColor }}
+          className="pw-cursor-pointer"
+          onClick={toggleTabsMemo}
+        />
       ) : (
         <HamburguerIcon
+          style={{ stroke: textColor }}
           onClick={toggleTabsMemo}
           className="pw-cursor-pointer"
         />
       )}
-      {opened && (
+      {opened ? (
         <div className="pw-flex pw-flex-col pw-bg-white pw-absolute pw-top-[90px] pw-left-0 pw-w-screen pw-z-30 pw-shadow-inner pw-py-8 pw-items-center pw-gap-y-4">
           {tabs?.map((tab) => (
             <a
+              style={{ color: textColor }}
               href={tab.router}
               className={`pw-font-montserrat pw-font-[600] pw-text-sm ${classNames?.tabClassName}`}
               key={tab.name}
@@ -52,23 +60,25 @@ export const NavigationTabsPixwaySDKMobile = ({
           {!session && (
             <div className="pw-flex pw-justify-center pw-gap-x-[26px]">
               <PixwayButton
-                onClick={() => router.push(signInRoute)}
+                onClick={() => router.pushConnect(signInRoute)}
                 fullWidth
                 className="!pw-bg-brand-primary !pw-px-[40px] !pw-text-white !pw-text-xs !pw-py-[9px] pw-rounded-[48px] pw-shadow-[0px_2px_4px_rgba(0,0,0,0.26)]"
               >
                 {translate('shared>login')}
               </PixwayButton>
-              <PixwayButton
-                onClick={() => router.push(signUpRoute)}
-                fullWidth
-                className="!pw-bg-[#EFEFEF] !pw-px-[40px] !pw-text-black !pw-text-xs !pw-py-[9px] pw-rounded-[48px]  !pw-border-[#DCDCDC] !pw-border-1"
-              >
-                {translate('shared>register')}
-              </PixwayButton>
+              {hasSignUp && (
+                <PixwayButton
+                  onClick={() => router.pushConnect(signUpRoute)}
+                  fullWidth
+                  className="!pw-bg-[#EFEFEF] !pw-px-[40px] !pw-text-black !pw-text-xs !pw-py-[9px] pw-rounded-[48px]  !pw-border-[#DCDCDC] !pw-border-1"
+                >
+                  {translate('shared>register')}
+                </PixwayButton>
+              )}
             </div>
           )}
         </div>
-      )}
+      ) : null}
     </div>
-  );
+  ) : null;
 };