+++ src/modules/shared/hooks/usePrivateRoute/usePrivateRoute.ts
@@ -5,7 +5,7 @@ import { UserRoleEnum } from '@w3block/sdk-id';
 import { PixwayAppRoutes } from '../../enums/PixwayAppRoutes';
 import { PrivateRouteStrategy } from '../../enums/PrivateRouteStrategy';
 import { usePixwaySession } from '../usePixwaySession';
-import useRouter from '../useRouter';
+import { useRouterConnect } from '../useRouterConnect';
 import { useSessionUser } from '../useSessionUser';
 
 interface Config {
@@ -18,17 +18,17 @@ export const usePrivateRoute = (
     roles: [],
     strategy: PrivateRouteStrategy.BLOCK_SPECIFIED_ROLES,
   },
-  redirectRoute: PixwayAppRoutes = PixwayAppRoutes.HOME
+  redirectRoute: PixwayAppRoutes = PixwayAppRoutes.SIGN_IN
 ) => {
   const { status } = usePixwaySession();
   const user = useSessionUser();
-  const router = useRouter();
+  const router = useRouterConnect();
   const { roles, strategy } = config ?? {};
   const isLoading = status !== 'authenticated';
 
   useEffect(() => {
     if (status === 'unauthenticated') {
-      router.push(redirectRoute);
+      router.pushConnect(redirectRoute);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [status === 'loading', redirectRoute]);