+++ src/modules/tokens/components/TokenDetailsTemplate/TokenDetailsTemplate.tsx
@@ -1,7 +1,9 @@
 import { InternalPagesLayoutBase } from '../../../shared';
 import { MintedInfoCard } from '../../../shared/components/MintedInfoCard';
 import TranslatableComponent from '../../../shared/components/TranslatableComponent';
-import useRouter from '../../../shared/hooks/useRouter';
+import { PixwayAppRoutes } from '../../../shared/enums/PixwayAppRoutes';
+import { useRouterConnect } from '../../../shared/hooks/useRouterConnect';
+import ChevronLeft from '../../assets/ArrowPagination.svg';
 import { usePublicTokenData } from '../../hooks/usePublicTokenData';
 import {
   Dimensions2DValue,
@@ -10,7 +12,7 @@ import {
 import { TokenDetailsCard } from '../TokenDetailsCard';
 
 const _TokenDetailsTemplate = () => {
-  const router = useRouter();
+  const router = useRouterConnect();
   const contractAddress = (router.query.contractAddress as string) ?? '';
   const chainId = (router.query.chainId as string) ?? '';
   const tokenId = (router.query.tokenId as string) ?? '';
@@ -22,6 +24,18 @@ const _TokenDetailsTemplate = () => {
 
   return publicTokenResponse ? (
     <div className="pw-w-full sm:pw-max-w-[968px] pw-font-roboto">
+      <div className="sm:pw-hidden pw-mb-4">
+        <div className="pw-flex pw-items-center pw-gap-x-4 pw-mb-[16.5px]">
+          <button
+            onClick={() => router.pushConnect(PixwayAppRoutes.TOKENS)}
+            className="pw-flex pw-items-center pw-justify-center pw-rounded-full pw-border pw-border-[#777E8F] pw-bg-[#F7F7F7] pw-w-4 pw-h-4 pw-ml-[22px]"
+          >
+            <ChevronLeft className="pw-w-[4.25px] pw-h-[8.5px] pw-fill-[#777E8F]" />
+          </button>
+        </div>
+
+        <div className="pw-w-full pw-h-[1px] pw-bg-[#EFEFEF]" />
+      </div>
       <TokenDetailsCard
         tokenData={
           (publicTokenResponse?.data?.dynamicInformation.tokenData as Record<